+++ src/App.tsx
@@ -1,0 +1,13 @@
+import MainLayout from './layouts/MainLayout.tsx'
+import HomePage from './pages/HomePage.tsx'
+
+function App() {
+  return (
+    <>
+      <MainLayout/>
+      <HomePage/>
+    </>
+  )
+}
+
+export default App
