+++ src/pages/HomePage.tsx
@@ -1,0 +1,8 @@
+
+export default function HomePage() {
+    return (
+      <>
+      </>
+    )
+}
+
